refactor: reorganize project structure and fix warnings (#41)
* feat: refactor resume generation and data structure for improved clarity and functionality

* refactor: simplify skill listing rules in resume generation

* Refactor resume components and update imports

- Changed import paths for MermaidRenderer and ResumeDetail components to relative paths.
- Added new ResumeBuilder, ResumeCard, ResumeDetail, ResumePDF, ResumePreview components for improved resume functionality.
- Implemented ResumeBuilder with state management for resume generation and history fetching.
- Created ResumeCard for displaying previously generated resumes.
- Developed ResumePDF for generating downloadable PDF resumes.
- Introduced ResumePreview for rendering resume previews.
- Removed unused StatsGrid and Footer components.
- Added HeroSection and SocialLinks components for enhanced homepage layout.
- Cleaned up layout and feature exports by removing unnecessary index files.

* feat: enhance SectionWrapper visibility logic and improve image handling in ProjectDetailPage
