+++ app/career/components/CareerCard.tsx
