+++ app/education/components/EducationCard.tsx
