+++ app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import { HeroSection } from "@/components/features";
+import HeroSection from "@/components/home/HeroSection";
 
 const Home: React.FC = () => {
   return <HeroSection />;
