+++ app/projects/(list)/components/GitHubStats.tsx
