+++ app/projects/(list)/components/ProjectCard.tsx
