--- components/cards/index.tsx
@@ -1,3 +1,0 @@
-export { default as CareerCard } from "./CareerCard";
-export { default as EducationCard } from "./EducationCard";
-export { default as ProjectCard } from "./ProjectCard";
