--- components/features/index.ts
@@ -1,8 +1,0 @@
-export { default as ToggleProfileQR } from "./ToggleProfileQR";
-export { default as SocialLinks } from "./SocialLinks";
-export { default as HeroSection } from "./HeroSection";
-export { default as StatsGrid } from "./StatsGrid";
-export { default as SectionWrapper } from "./SectionWrapper";
-export { default as ProjectsList } from "./ProjectsList";
-export { default as ProjectsFilterBar } from "./ProjectsFilterBar";
-export { default as ProjectsPagination } from "./ProjectsPagination";
