+++ components/home/HeroSection.tsx
