+++ components/home/ToggleProfileQR.tsx
