--- components/layout/index.ts
@@ -1,3 +1,0 @@
-export { default as Header } from "./Header";
-export { default as Footer } from "./Footer";
-export { default as MobileNav } from "./MobileNav";
