--- components/resume/index.ts
@@ -1,5 +1,0 @@
-export { default as ResumeBuilder } from "./ResumeBuilder";
-export { default as ResumePreview } from "./ResumePreview";
-export { default as ResumeCard } from "./ResumeCard";
-export { default as ResumeDetail } from "./ResumeDetail";
-export { ResumePDFLink } from "./ResumePDF";
