--- config/index.ts
@@ -1,7 +1,0 @@
-/**
- * Configuration Index
- * Re-export all configuration modules
- */
-
-export * from "./contact-info";
-export * from "./site-config";
